Massive commit... oops.
Beginning of the form generator and data entry module. Forms are
built from the form definitions using the FormManager singleton.

There will be no direct link between the templates and the forms
other than the Tetsable.datatype field.

Currently, subfields are not listening to any conditional display
requirements, and the entered data isn't saving. Nor have any
validation routines been created.

But, the forms are displaying well, and it should be good enough to
show to MTC tomorrow. eek